Move json results to different file
