+++ coppers_test_runner/test_runner/mod.rs
@@ -14,18 +14,17 @@
 
 // Note that this is heavily inspired by libtest that is part of the Rust language.
 
+use self::json::write_to_json;
 use coppers_sensors::{RAPLSensor, Sensor};
-use git2::Repository;
 use std::any::Any;
-use std::env::current_dir;
-use std::fs::File;
 use std::io::{self, Write};
 use std::panic::catch_unwind;
 use std::sync::{Arc, Mutex};
-use std::time::SystemTime;
 use test::{StaticTestFn, TestDescAndFn};
 
-const REPEAT_TESTS_AMOUNT_OF_TIMES: usize = 1;
+mod json;
+
+pub(crate) const REPEAT_TESTS_AMOUNT_OF_TIMES: usize = 1;
 
 pub fn runner(tests: &[&test::TestDescAndFn]) {
     let tests: Vec<_> = tests.iter().map(make_owned_test).collect();
@@ -73,38 +72,6 @@ pub fn runner(tests: &[&test::TestDescAndFn]) {
     println!("test result: {}.\n\t{} passed;\n\t{} failed;\n\t{ignored} ignored;\n\tfinished in {total_us} μs consuming {total_uj} μJ\n\tspend {test_us} μs and {test_uj} μJ on tests\n\tspend {overhead_us} μs and {overhead_uj} μJ on overhead", passed(failed_tests.is_empty()), passed_tests.len(), failed_tests.len());
     // Write test results to JSON file
     write_to_json(passed_tests, total_us, total_uj, overhead_us, overhead_uj)
-}
-
-fn write_to_json(
-    passed_tests: Vec<CompletedTest>,
-    total_us: u128,
-    total_uj: u128,
-    overhead_us: u128,
-    overhead_uj: u128,
-) {
-    // Get git hash of last commit
-    // WARNING: ASSUMING REPO EXISTS!
-    let current_directory = current_dir().unwrap();
-    let current_path = current_directory.as_path().to_str().unwrap();
-    let repo = Repository::open(current_path).unwrap();
-    let head_hash = repo.head().unwrap().target().unwrap();
-    let head_hash_bytes = head_hash.as_bytes();
-    let head_hash_json = serde_json::to_string(&head_hash_bytes).unwrap();
-
-    // Get the timestamp of the current time
-    let timestamp = SystemTime::now()
-        .duration_since(SystemTime::UNIX_EPOCH)
-        .unwrap()
-        .as_secs();
-
-    // Convert test results in JSON object
-    let json_tests = serde_json::to_string(&passed_tests).unwrap();
-    let full_json = format!("{{\"timestamp\":{timestamp},\"head\":{head_hash_json},\"total_time\":{total_us},\"total_consumption\":{total_uj},\"overhead_time\":{overhead_us},\"overhead_consumption\":{overhead_uj},\"number_of_repeats\":{REPEAT_TESTS_AMOUNT_OF_TIMES},\"tests\":{json_tests}}}");
-
-    let json_file_name = format!("target/coppers_results-{timestamp}.json");
-    let mut file = File::create(json_file_name).unwrap();
-    file.write_all(full_json.as_bytes())
-        .expect("Writing of JSON file failed.");
 }
 
 fn print_failures(tests: &Vec<CompletedTest>) -> std::io::Result<()> {
@@ -168,7 +135,7 @@ enum TestResult {
 }
 
 #[derive(serde::Serialize, Debug, PartialEq)]
-struct CompletedTest {
+pub(crate) struct CompletedTest {
     name: String,
     state: TestResult,
     uj: Option<u128>,
